Move Nullable back to font-types crate
I think this is a better separation of concerns: font-types contains the
type definitions, but the logic for resolution etc lives in font-tables.
